Rename VectorProduct to CreateMask and move implementations.
Rename SimdVectorProduct/Simd::VectorProduct to SimdCreateMask/Simd::CreateMask,
move architecture implementations into dedicated CreateMask.cpp files, relocate
the auto-test to TestDrawing.cpp, update VS2022 projects, and refresh release
7.2.165 notes in docs/2026.html.
