+++ prj/vs2022/Avx2.vcxproj
@@ -70,6 +70,7 @@
     <ClCompile Include="..\..\src\Simd\SimdAvx2MidpointFilter.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx2Neural.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx2NeuralConvolution.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdAvx2CreateMask.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx2Operation.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx2RecursiveBilateralFilter.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx2Reduce.cpp" />
