+++ prj/vs2022/Avx512bw.vcxproj
@@ -71,6 +71,7 @@
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwMidpointFilter.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwNeural.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwNeuralConvolution.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdAvx512bwCreateMask.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwOperation.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwReduce.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwReduceGray2x2.cpp" />
