+++ prj/vs2022/Avx512bw.vcxproj.filters
@@ -424,6 +424,9 @@
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwInterleave.cpp">
       <Filter>Avx512bw\Convert</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Simd\SimdAvx512bwCreateMask.cpp">
+      <Filter>Avx512bw\Image</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\Simd\SimdAvx512bwFill.cpp">
       <Filter>Avx512bw\Image</Filter>
     </ClCompile>
