+++ prj/vs2022/Sse41.vcxproj
@@ -74,6 +74,7 @@
     <ClCompile Include="..\..\src\Simd\SimdSse41MidpointFilter.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSse41Neural.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSse41NeuralConvolution.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdSse41CreateMask.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSse41Operation.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSse41RecursiveBilateralFilter.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSse41Reduce.cpp" />
