+++ prj/vs2022/Sve2.vcxproj
@@ -79,6 +79,7 @@
     <ClCompile Include="..\..\src\Simd\SimdSve2MinFilter.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSve2Neural.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSve2NeuralConvolution.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdSve2CreateMask.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSve2Operation.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSve2Reduce.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSve2ReduceGray2x2.cpp" />
