+++ prj/vs2022/Sve2.vcxproj.filters
@@ -278,6 +278,9 @@
     <ClCompile Include="..\..\src\Simd\SimdSve2InterleaveUv.cpp">
       <Filter>Sve2\Convert</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Simd\SimdSve2CreateMask.cpp">
+      <Filter>Sve2\Image</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\Simd\SimdSve2Fill.cpp">
       <Filter>Sve2\Image</Filter>
     </ClCompile>
