+++ src/Simd/SimdAvx2Operation.cpp
@@ -23,7 +23,6 @@
 */
 #include "Simd/SimdMemory.h"
 #include "Simd/SimdStore.h"
-#include "Simd/SimdAlphaBlending.h"
 
 namespace Simd
 {
@@ -191,40 +190,6 @@ namespace Simd
                 OperationBinary16i<true>(a, aStride, b, bStride, width, height, dst, dstStride, type);
             else
                 OperationBinary16i<false>(a, aStride, b, bStride, width, height, dst, dstStride, type);
-        }
-
-        template <bool align> SIMD_INLINE void VectorProduct(const __m256i & vertical, const uint8_t * horizontal, uint8_t * dst)
-        {
-            __m256i _horizontal = Load<align>((__m256i*)horizontal);
-            __m256i lo = Divide16uBy255(_mm256_mullo_epi16(vertical, _mm256_unpacklo_epi8(_horizontal, K_ZERO)));
-            __m256i hi = Divide16uBy255(_mm256_mullo_epi16(vertical, _mm256_unpackhi_epi8(_horizontal, K_ZERO)));
-            Store<align>((__m256i*)dst, _mm256_packus_epi16(lo, hi));
-        }
-
-        template <bool align> void VectorProduct(const uint8_t * vertical, const uint8_t * horizontal, uint8_t * dst, size_t stride, size_t width, size_t height)
-        {
-            assert(width >= A);
-            if (align)
-                assert(Aligned(horizontal) && Aligned(dst) && Aligned(stride));
-
-            size_t alignedWidth = Simd::AlignLo(width, A);
-            for (size_t row = 0; row < height; ++row)
-            {
-                __m256i _vertical = _mm256_set1_epi16(vertical[row]);
-                for (size_t col = 0; col < alignedWidth; col += A)
-                    VectorProduct<align>(_vertical, horizontal + col, dst + col);
-                if (alignedWidth != width)
-                    VectorProduct<false>(_vertical, horizontal + width - A, dst + width - A);
-                dst += stride;
-            }
-        }
-
-        void VectorProduct(const uint8_t * vertical, const uint8_t * horizontal, uint8_t * dst, size_t stride, size_t width, size_t height)
-        {
-            if (Aligned(horizontal) && Aligned(dst) && Aligned(stride))
-                VectorProduct<true>(vertical, horizontal, dst, stride, width, height);
-            else
-                VectorProduct<false>(vertical, horizontal, dst, stride, width, height);
         }
     }
 #endif// SIMD_AVX2_ENABLE
