+++ src/Simd/SimdSve2Operation.cpp
@@ -176,47 +176,6 @@ namespace Simd
                 assert(0);
             }
         }
-
-        //-------------------------------------------------------------------------------------------------
-
-        SIMD_INLINE svuint8_t DivideI16By255(const svuint16_t & value)
-        {
-            const svbool_t full = svptrue_b16();
-            svuint16_t sum = svadd_n_u16_x(full, value, 1);
-            sum = svadd_u16_x(full, sum, svlsr_n_u16_x(full, sum, 8));
-            return svqxtnb_u16(svlsr_n_u16_x(full, sum, 8));
-        }
-
-        SIMD_INLINE svuint8_t VectorProduct(const svuint16_t & vertical, const svuint8_t & horizontal)
-        {
-            const svbool_t full = svptrue_b16();
-            svuint8_t lo = DivideI16By255(svmul_u16_x(full, vertical, svunpklo_u16(horizontal)));
-            svuint8_t hi = DivideI16By255(svmul_u16_x(full, vertical, svunpkhi_u16(horizontal)));
-            return svuzp1_u8(lo, hi);
-        }
-
-        void VectorProduct(const uint8_t * vertical, const uint8_t * horizontal, uint8_t * dst, size_t stride, size_t width, size_t height)
-        {
-            size_t A = svcntb();
-            size_t widthA = AlignLo(width, A);
-            for (size_t row = 0; row < height; ++row)
-            {
-                const svuint16_t _vertical = svdup_n_u16(vertical[row]);
-                size_t col = 0;
-                for (; col < widthA; col += A)
-                {
-                    svuint8_t _horizontal = svld1_u8(svptrue_b8(), horizontal + col);
-                    svst1_u8(svptrue_b8(), dst + col, VectorProduct(_vertical, _horizontal));
-                }
-                if (widthA < width)
-                {
-                    svbool_t tail = svwhilelt_b8(col, width);
-                    svuint8_t _horizontal = svld1_u8(tail, horizontal + col);
-                    svst1_u8(tail, dst + col, VectorProduct(_vertical, _horizontal));
-                }
-                dst += stride;
-            }
-        }
     }
 #endif
 }
