+++ src/Test/TestOperation.cpp
@@ -270,114 +270,4 @@ namespace Test
 
         return result;
     }
-
-    //-------------------------------------------------------------------------------------------------
-
-#if defined(_MSC_VER)
-#pragma warning (push)
-#pragma warning (disable: 4996)
-#endif
-
-#if defined(__GNUC__)
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wdeprecated-declarations"
-#endif
-
-    namespace 
-    {
-        struct FuncVP
-        {
-            typedef void(*FuncPtr)(const uint8_t* vertical, const uint8_t* horizontal, uint8_t* dst, size_t stride, size_t width, size_t height);
-
-            FuncPtr func;
-            String description;
-
-            FuncVP(const FuncPtr& f, const String& d) : func(f), description(d) {}
-
-            void Call(const View& v, const View& h, View& dst) const
-            {
-                TEST_PERFORMANCE_TEST(description);
-                func(v.data, h.data, dst.data, dst.stride, dst.width, dst.height);
-            }
-        };
-    }
-
-#define ARGS_VP(function) FuncVP(function, std::string(#function))
-
-    bool VectorProductAutoTest(int width, int height, const FuncVP & f1, const FuncVP & f2)
-    {
-        bool result = true;
-
-        TEST_LOG_SS(Info, "Test " << f1.description << " & " << f2.description << " [" << width << ", " << height << "].");
-
-        View v(height, 1, View::Gray8, NULL, TEST_ALIGN(height));
-        FillRandom(v);
-        View h(width, 1, View::Gray8, NULL, TEST_ALIGN(width));
-        FillRandom(h);
-
-        View d1(width, height, View::Gray8, NULL, TEST_ALIGN(width));
-        View d2(width, height, View::Gray8, NULL, TEST_ALIGN(width));
-
-        TEST_EXECUTE_AT_LEAST_MIN_TIME(f1.Call(v, h, d1));
-
-        TEST_EXECUTE_AT_LEAST_MIN_TIME(f2.Call(v, h, d2));
-
-        result = result && Compare(d1, d2, 0, true, 64);
-
-        return result;
-    }
-
-    bool VectorProductAutoTest(const FuncVP & f1, const FuncVP & f2)
-    {
-        bool result = true;
-
-        result = result && VectorProductAutoTest(W, H, f1, f2);
-        result = result && VectorProductAutoTest(W - O, H + O, f1, f2);
-
-        return result;
-    }
-
-    bool VectorProductAutoTest(const Options & options)
-    {
-        bool result = true;
-
-        if (TestBase(options))
-            result = result && VectorProductAutoTest(ARGS_VP(Simd::Base::VectorProduct), ARGS_VP(SimdVectorProduct));
-
-#ifdef SIMD_SSE41_ENABLE
-        if (Simd::Sse41::Enable && TestSse41(options))
-            result = result && VectorProductAutoTest(ARGS_VP(Simd::Sse41::VectorProduct), ARGS_VP(SimdVectorProduct));
-#endif 
-
-#ifdef SIMD_AVX2_ENABLE
-        if (Simd::Avx2::Enable && TestAvx2(options))
-            result = result && VectorProductAutoTest(ARGS_VP(Simd::Avx2::VectorProduct), ARGS_VP(SimdVectorProduct));
-#endif 
-
-#ifdef SIMD_AVX512BW_ENABLE
-        if (Simd::Avx512bw::Enable && TestAvx512bw(options))
-            result = result && VectorProductAutoTest(ARGS_VP(Simd::Avx512bw::VectorProduct), ARGS_VP(SimdVectorProduct));
-#endif 
-
-#ifdef SIMD_NEON_ENABLE
-        if (Simd::Neon::Enable && TestNeon(options))
-            result = result && VectorProductAutoTest(ARGS_VP(Simd::Neon::VectorProduct), ARGS_VP(SimdVectorProduct));
-#endif 
-
-#ifdef SIMD_SVE2_ENABLE
-        if (Simd::Sve2::Enable && TestSve2(options))
-            result = result && VectorProductAutoTest(ARGS_VP(Simd::Sve2::VectorProduct), ARGS_VP(SimdVectorProduct));
-#endif 
-
-        return result;
-    }
-
-#if defined(_MSC_VER)
-#pragma warning (pop)
-#endif
-
-#if defined(__GNUC__)
-#pragma GCC diagnostic pop
-#endif
-
 }
